requirements.txt
Package                       Version
----------------------------- ---------
asgiref                       3.8.1
attrs                         25.1.0
boolean.py                    5.0
CacheControl                  0.14.2
certifi                       2025.1.31
charset-normalizer            3.4.1
colorama                      0.4.6
cyclonedx-python-lib          9.1.0
defusedxml                    0.7.1
Django                        5.1.6
django-environ                0.12.0
djangorestframework           3.15.2
djangorestframework_simplejwt 5.5.0
drf-spectacular               0.28.0
exceptiongroup                1.2.2
filelock                      3.18.0
idna                          3.10
inflection                    0.5.1
iniconfig                     2.1.0
jsonschema                    4.23.0
jsonschema-specifications     2024.10.1
license-expression            30.4.1
markdown-it-py                3.0.0
mdurl                         0.1.2
msgpack                       1.1.0
packageurl-python             0.16.0
packaging                     25.0
pip                           25.0.1
pip-api                       0.0.34
pip_audit                     2.9.0
pip-requirements-parser       32.0.1
platformdirs                  4.3.7
pluggy                        1.5.0
psycopg2-binary               2.9.10
py-serializable               2.0.0
Pygments                      2.19.1
PyJWT                         2.9.0
pyparsing                     3.2.3
pytest                        8.3.5
pytest-django                 4.11.1
PyYAML                        6.0.2
referencing                   0.36.2
requests                      2.32.3
rich                          14.0.0
rpds-py                       0.22.3
setuptools                    65.5.1
sortedcontainers              2.4.0
sqlparse                      0.5.3
toml                          0.10.2
tomli                         2.2.1
typing_extensions             4.12.2
tzdata                        2025.1
uritemplate                   4.1.1
urllib3                       2.4.0
whitenoise                    6.9.0
